add helpers to base for configure and import
