+++ internal/vault/sys/password_policy.go
@@ -11,22 +11,21 @@ import (
 
 	"github.com/hashicorp/terraform-provider-vault/internal/framework/base"
 	"github.com/hashicorp/terraform-provider-vault/internal/framework/client"
-	"github.com/hashicorp/terraform-provider-vault/internal/provider"
 )
 
-// ensure we our interface implementation is correct
+// Ensure the implementation satisfies the resource.ResourceWithConfigure interface
 var _ resource.ResourceWithConfigure = &PasswordPolicyResource{}
 
+// NewPasswordPolicyResource returns the implementation for this resource to be
+// imported by the Terraform Plugin Framework provider
 func NewPasswordPolicyResource() resource.Resource {
 	return &PasswordPolicyResource{}
 }
 
+// PasswordPolicyResource implements the methods that define this resource
 type PasswordPolicyResource struct {
-	meta *provider.ProviderMeta
-}
-
-func (r *PasswordPolicyResource) Metadata(_ context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
-	resp.TypeName = req.ProviderTypeName + "_password_policy"
+	base.ResourceWithConfigure
+	base.WithImportByID
 }
 
 // PasswordPolicyModel describes the Terraform resource data model to match the
@@ -48,21 +47,17 @@ type PasswordPolicyAPIModel struct {
 	Policy string `json:"policy" mapstructure:"policy"`
 }
 
-func (r *PasswordPolicyResource) Configure(_ context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
-	if req.ProviderData == nil {
-		return
-	}
-	meta, ok := req.ProviderData.(*provider.ProviderMeta)
-	if !ok {
-		resp.Diagnostics.AddError(
-			"Unexpected Resource Configure Type",
-			fmt.Sprintf("Expected *provider.ProviderMeta, got: %T. Please report this issue to the provider developers.", req.ProviderData),
-		)
-		return
-	}
-	r.meta = meta
+// Metadata defines the resource name as it would appear in Terraform configurations
+//
+// https://developer.hashicorp.com/terraform/plugin/framework/resources#metadata-method
+func (r *PasswordPolicyResource) Metadata(_ context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_password_policy"
 }
 
+// Schema defines this resource's schema which is the data that is available in
+// the resource's configuration, plan, and state
+//
+// https://developer.hashicorp.com/terraform/plugin/framework/resources#schema-method
 func (r *PasswordPolicyResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		Attributes: map[string]schema.Attribute{
@@ -81,6 +76,9 @@ func (r *PasswordPolicyResource) Schema(ctx context.Context, req resource.Schema
 	base.MustAddLegacyBaseSchema(&resp.Schema)
 }
 
+// Create is called during the terraform apply command.
+//
+// https://developer.hashicorp.com/terraform/plugin/framework/resources/create
 func (r *PasswordPolicyResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
 	var plan PasswordPolicyModel
 
@@ -91,7 +89,7 @@ func (r *PasswordPolicyResource) Create(ctx context.Context, req resource.Create
 		return
 	}
 
-	client, err := client.GetClient(ctx, r.meta, plan.Namespace.ValueString())
+	client, err := client.GetClient(ctx, r.Meta(), plan.Namespace.ValueString())
 	if err != nil {
 		resp.Diagnostics.AddError("Error Configuring Resource Client", err.Error())
 		return
@@ -115,12 +113,16 @@ func (r *PasswordPolicyResource) Create(ctx context.Context, req resource.Create
 	}
 
 	// write the ID to state which is required for acceptance testing
-	plan.ID = types.StringValue(path)
+	plan.ID = types.StringValue(plan.Name.ValueString())
 
 	// Save data into Terraform state
 	resp.Diagnostics.Append(resp.State.Set(ctx, &plan)...)
 }
 
+// Read is called during the terraform apply, terraform plan, and terraform
+// refresh commands.
+//
+// https://developer.hashicorp.com/terraform/plugin/framework/resources/read
 func (r *PasswordPolicyResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
 	var state PasswordPolicyModel
 	// Read Terraform prior state data into the model
@@ -130,7 +132,7 @@ func (r *PasswordPolicyResource) Read(ctx context.Context, req resource.ReadRequ
 		return
 	}
 
-	client, err := client.GetClient(ctx, r.meta, state.Namespace.ValueString())
+	client, err := client.GetClient(ctx, r.Meta(), state.Namespace.ValueString())
 	if err != nil {
 		resp.Diagnostics.AddError("Error Configuring Resource Client", err.Error())
 		return
@@ -186,11 +188,14 @@ func (r *PasswordPolicyResource) Read(ctx context.Context, req resource.ReadRequ
 	state.Policy = types.StringValue(readResp.Policy)
 
 	// write the ID to state which is required for acceptance testing
-	state.ID = types.StringValue(path)
+	state.ID = types.StringValue(state.Name.ValueString())
 
 	resp.Diagnostics.Append(resp.State.Set(ctx, &state)...)
 }
 
+// Update is called during the terraform apply command
+//
+// https://developer.hashicorp.com/terraform/plugin/framework/resources/update
 func (r *PasswordPolicyResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
 	var plan PasswordPolicyModel
 
@@ -201,7 +206,7 @@ func (r *PasswordPolicyResource) Update(ctx context.Context, req resource.Update
 		return
 	}
 
-	client, err := client.GetClient(ctx, r.meta, plan.Namespace.ValueString())
+	client, err := client.GetClient(ctx, r.Meta(), plan.Namespace.ValueString())
 	if err != nil {
 		resp.Diagnostics.AddError("Error Configuring Resource Client", err.Error())
 		return
@@ -225,12 +230,15 @@ func (r *PasswordPolicyResource) Update(ctx context.Context, req resource.Update
 	}
 
 	// write the ID to state which is required for acceptance testing
-	plan.ID = types.StringValue(path)
+	plan.ID = types.StringValue(plan.Name.ValueString())
 
 	// Save data into Terraform state
 	resp.Diagnostics.Append(resp.State.Set(ctx, &plan)...)
 }
 
+// Delete is called during the terraform apply command
+//
+// https://developer.hashicorp.com/terraform/plugin/framework/resources/delete
 func (r *PasswordPolicyResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
 	var plan PasswordPolicyModel
 
@@ -241,7 +249,7 @@ func (r *PasswordPolicyResource) Delete(ctx context.Context, req resource.Delete
 		return
 	}
 
-	client, err := client.GetClient(ctx, r.meta, plan.Namespace.ValueString())
+	client, err := client.GetClient(ctx, r.Meta(), plan.Namespace.ValueString())
 	if err != nil {
 		resp.Diagnostics.AddError("Error Configuring Resource Client", err.Error())
 		return
